+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # 1.20.1 Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
+## [Unreleased]
+
+## [1.3.0]
+### Added
+ - Programmed Circuit Card now works with Pattern Buffers. (by @Cyrenee)
+
+## [1.2.11]
+### Added
+ - Expanded AE's Smart Blocking takes into account differences in Programmed Circuits. (by @Zatone0) [#56](https://github.com/yuuki1293/ProgrammedCircuitCard/issues/56)
+
 ## [1.2.10]
 ### Added
  - Supports Giga Pattern Provider (Expanded AE).